feature : adding view page
category controller
View using thymeleaf, fetch
