+++ src/main/java/com/chatboard/etude/controller/category/CategoryController.java
@@ -1,49 +1,25 @@
 package com.chatboard.etude.controller.category;
 
-import com.chatboard.etude.dto.category.CategoryCreateRequest;
-import com.chatboard.etude.dto.response.Response;
-import com.chatboard.etude.entity.category.Category;
 import com.chatboard.etude.service.category.CategoryService;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.servlet.ModelAndView;
 
-import javax.validation.Valid;
-
-@Api(value = "CategoryController", tags = "Category")
-@RestController
+@Controller
 @RequiredArgsConstructor
 public class CategoryController {
 
     private final CategoryService categoryService;
 
-    @ApiOperation(value = "read all category",
-            notes = "This request reads all categories.")
-    @GetMapping("api/categories")
+    @GetMapping("/categories")
     @ResponseStatus(HttpStatus.OK)
-    public Response readAll() {
-        return Response.success(categoryService.readAll());
+    public ModelAndView readAll() {
+        ModelAndView modelAndView = new ModelAndView("/api/categories");
+        modelAndView.addObject(categoryService.readAll());
+        return modelAndView;
     }
 
-    @ApiOperation(value = "create categories",
-            notes = "This request create a category.")
-    @PostMapping("api/categories")
-    @ResponseStatus(HttpStatus.CREATED)
-    public Response create(@Valid @RequestBody CategoryCreateRequest request) {
-        categoryService.create(request);
-        return Response.success();
-    }
-
-    @ApiOperation(value = "delete category",
-            notes = "This request deletes category.")
-    @DeleteMapping("/api/categories/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public Response delete(@ApiParam(value = "category id", required = true)
-                           @PathVariable Long id) {
-        categoryService.delete(id);
-        return Response.success();
-    }
 }
